Add Dockerfile
Had to use a shim bash script to map the env vars to java properties.

Unfortunate that typesafe config doesn't support env vars directly.

The default JAVA_OPTS are totally a guess.

This patch assumes my TCP_CONNECT_TIMEOUT patch, but I can rework if the former
is not desired.